requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
qdrant-client==1.9.1
httpx==0.27.0
python-multipart==0.0.9
pypdf==4.2.0
pytest==8.2.2
python-dotenv==1.0.1
